+++ otelconf/x/config_json.go
@@ -123,6 +123,21 @@ func (j *TraceContextPropagator) UnmarshalJSON(b []byte) error {
 }
 
 // UnmarshalJSON implements json.Unmarshaler.
+func (j *ExperimentalGCPResourceDetector) UnmarshalJSON(b []byte) error {
+	type plain ExperimentalGCPResourceDetector
+	var p plain
+	if err := json.Unmarshal(b, &p); err != nil {
+		return errors.Join(newErrUnmarshal(j), err)
+	}
+	if p == nil {
+		*j = ExperimentalGCPResourceDetector{}
+	} else {
+		*j = ExperimentalGCPResourceDetector(p)
+	}
+	return nil
+}
+
+// UnmarshalJSON implements json.Unmarshaler.
 func (j *ExperimentalAWSECSResourceDetector) UnmarshalJSON(b []byte) error {
 	type plain ExperimentalAWSECSResourceDetector
 	var p plain
@@ -148,6 +163,21 @@ func (j *ExperimentalAWSEKSResourceDetector) UnmarshalJSON(b []byte) error {
 		*j = ExperimentalAWSEKSResourceDetector{}
 	} else {
 		*j = ExperimentalAWSEKSResourceDetector(p)
+	}
+	return nil
+}
+
+// UnmarshalJSON implements json.Unmarshaler.
+func (j *ExperimentalAzureVMResourceDetector) UnmarshalJSON(b []byte) error {
+	type plain ExperimentalAzureVMResourceDetector
+	var p plain
+	if err := json.Unmarshal(b, &p); err != nil {
+		return errors.Join(newErrUnmarshal(j), err)
+	}
+	if p == nil {
+		*j = ExperimentalAzureVMResourceDetector{}
+	} else {
+		*j = ExperimentalAzureVMResourceDetector(p)
 	}
 	return nil
 }
@@ -222,8 +252,10 @@ func (j *ExperimentalResourceDetector) UnmarshalJSON(b []byte) error {
 	type Plain ExperimentalResourceDetector
 	type shadow struct {
 		Plain
+		GCP       json.RawMessage `json:"gcp"`
 		AWSECS    json.RawMessage `json:"aws.ecs"`
 		AWSEKS    json.RawMessage `json:"aws.eks"`
+		AzureVM   json.RawMessage `json:"azure.vm"`
 		Container json.RawMessage `json:"container"`
 		Host      json.RawMessage `json:"host"`
 		Process   json.RawMessage `json:"process"`
@@ -232,6 +264,17 @@ func (j *ExperimentalResourceDetector) UnmarshalJSON(b []byte) error {
 	var sh shadow
 	if err := json.Unmarshal(b, &sh); err != nil {
 		return errors.Join(newErrUnmarshal(j), err)
+	}
+
+	if sh.GCP != nil {
+		var c ExperimentalGCPResourceDetector
+		if err := json.Unmarshal(sh.GCP, &c); err != nil {
+			return errors.Join(newErrUnmarshal(j), err)
+		}
+		if c == nil {
+			c = ExperimentalGCPResourceDetector{}
+		}
+		sh.Plain.GCP = c
 	}
 
 	if sh.AWSECS != nil {
@@ -254,6 +297,17 @@ func (j *ExperimentalResourceDetector) UnmarshalJSON(b []byte) error {
 			c = ExperimentalAWSEKSResourceDetector{}
 		}
 		sh.Plain.AWSEKS = c
+	}
+
+	if sh.AzureVM != nil {
+		var c ExperimentalAzureVMResourceDetector
+		if err := json.Unmarshal(sh.AzureVM, &c); err != nil {
+			return errors.Join(newErrUnmarshal(j), err)
+		}
+		if c == nil {
+			c = ExperimentalAzureVMResourceDetector{}
+		}
+		sh.Plain.AzureVM = c
 	}
 
 	if sh.Container != nil {