+++ otelconf/x/config_test.go
@@ -2284,16 +2284,22 @@ func TestUnmarshalResourceJson(t *testing.T) {
 	}{
 		{
 			name:       "valid with all detectors",
-			jsonConfig: []byte(`{"detection/development": {"detectors": [{"aws.ecs": null},{"aws.eks": null},{"container": null},{"host": null},{"process": null},{"service": null}]}}`),
-			yamlConfig: []byte("detection/development:\n  detectors:\n    - aws.ecs:\n    - aws.eks:\n    - container:\n    - host:\n    - process:\n    - service:"),
+			jsonConfig: []byte(`{"detection/development": {"detectors": [{"gcp": null},{"aws.ecs": null},{"aws.eks": null},{"azure.vm": null},{"container": null},{"host": null},{"process": null},{"service": null}]}}`),
+			yamlConfig: []byte("detection/development:\n  detectors:\n    - gcp:\n    - aws.ecs:\n    - aws.eks:\n    - azure.vm:\n    - container:\n    - host:\n    - process:\n    - service:"),
 			wantResource: Resource{
 				DetectionDevelopment: &ExperimentalResourceDetection{
 					Detectors: []ExperimentalResourceDetector{
+						{
+							GCP: ExperimentalGCPResourceDetector{},
+						},
 						{
 							AWSECS: ExperimentalAWSECSResourceDetector{},
 						},
 						{
 							AWSEKS: ExperimentalAWSEKSResourceDetector{},
+						},
+						{
+							AzureVM: ExperimentalAzureVMResourceDetector{},
 						},
 						{
 							Container: ExperimentalContainerResourceDetector{},
@@ -2313,16 +2319,22 @@ func TestUnmarshalResourceJson(t *testing.T) {
 		},
 		{
 			name:       "valid non-nil with all detectors",
-			jsonConfig: []byte(`{"detection/development": {"detectors": [{"aws.ecs": {}},{"aws.eks": {}},{"container": {}},{"host": {}},{"process": {}},{"service": {}}]}}`),
-			yamlConfig: []byte("detection/development:\n  detectors:\n    - aws.ecs: {}\n    - aws.eks: {}\n    - container: {}\n    - host: {}\n    - process: {}\n    - service: {}"),
+			jsonConfig: []byte(`{"detection/development": {"detectors": [{"gcp": {}},{"aws.ecs": {}},{"aws.eks": {}},{"azure.vm": {}},{"container": {}},{"host": {}},{"process": {}},{"service": {}}]}}`),
+			yamlConfig: []byte("detection/development:\n  detectors:\n    - gcp: {}\n    - aws.ecs: {}\n    - aws.eks: {}\n    - azure.vm: {}\n    - container: {}\n    - host: {}\n    - process: {}\n    - service: {}"),
 			wantResource: Resource{
 				DetectionDevelopment: &ExperimentalResourceDetection{
 					Detectors: []ExperimentalResourceDetector{
+						{
+							GCP: ExperimentalGCPResourceDetector{},
+						},
 						{
 							AWSECS: ExperimentalAWSECSResourceDetector{},
 						},
 						{
 							AWSEKS: ExperimentalAWSEKSResourceDetector{},
+						},
+						{
+							AzureVM: ExperimentalAzureVMResourceDetector{},
 						},
 						{
 							Container: ExperimentalContainerResourceDetector{},
@@ -2341,6 +2353,19 @@ func TestUnmarshalResourceJson(t *testing.T) {
 			},
 		},
 		{
+			name:       "invalid gcp detector",
+			jsonConfig: []byte(`{"detection/development": {"detectors": [{"gcp": 1}]}}`),
+			yamlConfig: []byte("detection/development:\n  detectors:\n    - gcp: 1"),
+			wantResource: Resource{
+				DetectionDevelopment: &ExperimentalResourceDetection{
+					Detectors: []ExperimentalResourceDetector{
+						{},
+					},
+				},
+			},
+			wantErrT: newErrUnmarshal(&ExperimentalResourceDetector{}),
+		},
+		{
 			name:       "invalid aws ecs detector",
 			jsonConfig: []byte(`{"detection/development": {"detectors": [{"aws.ecs": 1}]}}`),
 			yamlConfig: []byte("detection/development:\n  detectors:\n    - aws.ecs: 1"),
@@ -2357,6 +2382,19 @@ func TestUnmarshalResourceJson(t *testing.T) {
 			name:       "invalid aws eks detector",
 			jsonConfig: []byte(`{"detection/development": {"detectors": [{"aws.eks": 1}]}}`),
 			yamlConfig: []byte("detection/development:\n  detectors:\n    - aws.eks: 1"),
+			wantResource: Resource{
+				DetectionDevelopment: &ExperimentalResourceDetection{
+					Detectors: []ExperimentalResourceDetector{
+						{},
+					},
+				},
+			},
+			wantErrT: newErrUnmarshal(&ExperimentalResourceDetector{}),
+		},
+		{
+			name:       "invalid azure vm detector",
+			jsonConfig: []byte(`{"detection/development": {"detectors": [{"azure.vm": 1}]}}`),
+			yamlConfig: []byte("detection/development:\n  detectors:\n    - azure.vm: 1"),
 			wantResource: Resource{
 				DetectionDevelopment: &ExperimentalResourceDetection{
 					Detectors: []ExperimentalResourceDetector{