+++ otelconf/x/generated_config.go
@@ -976,6 +976,11 @@ type ExperimentalResourceDetection struct {
 }
 
 type ExperimentalResourceDetector struct {
+	// Enable the GCP resource detector.
+	// If omitted, ignore.
+	//
+	GCP ExperimentalGCPResourceDetector `json:"gcp,omitempty,omitzero" yaml:"gcp,omitempty" mapstructure:"gcp,omitempty"`
+
 	// Enable the AWS ECS resource detector.
 	// If omitted, ignore.
 	//
@@ -985,6 +990,11 @@ type ExperimentalResourceDetector struct {
 	// If omitted, ignore.
 	//
 	AWSEKS ExperimentalAWSEKSResourceDetector `json:"aws.eks,omitempty,omitzero" yaml:"aws.eks,omitempty" mapstructure:"aws.eks,omitempty"`
+
+	// Enable the Azure VM resource detector.
+	// If omitted, ignore.
+	//
+	AzureVM ExperimentalAzureVMResourceDetector `json:"azure.vm,omitempty,omitzero" yaml:"azure.vm,omitempty" mapstructure:"azure.vm,omitempty"`
 
 	// Enable the container resource detector, which populates container.* attributes.
 	// If omitted, ignore.
@@ -1010,9 +1020,13 @@ type ExperimentalResourceDetector struct {
 	AdditionalProperties interface{} `mapstructure:",remain"`
 }
 
+type ExperimentalGCPResourceDetector map[string]interface{}
+
 type ExperimentalAWSECSResourceDetector map[string]interface{}
 
 type ExperimentalAWSEKSResourceDetector map[string]interface{}
+
+type ExperimentalAzureVMResourceDetector map[string]interface{}
 
 type ExperimentalServiceResourceDetector map[string]interface{}